Implement demo command enhancements (issues #344-350)
Closes #343

This commit implements all subtasks for the demo command spike:

- #344: Promote `ampere demo` to top-level command by merging JazzDemoCommand into DemoCommand and registering it in Main.kt

- #345: Add timing profile flags (--fast, --detailed) with new DemoTiming.kt containing FAST (~45s), DEFAULT (~85s), and DETAILED (~120s) profiles

- #346: Add --quiet mode for CI/scripting with non-TUI output and machine-parseable format

- #347: Add --no-escalation flag to skip escalation prompts and use default choice (A) automatically

- #348: Refine escalation question wording to "Implementation scope:" with clearer option descriptions

- #349: Add golden output fallback for LLM failures via new GoldenOutput.kt containing pre-validated ObservabilitySpark.kt

- #350: Write comprehensive timing verification tests in DemoTimingTest.kt and GoldenOutputTest.kt

All tests pass. Demo supports both interactive TUI and quiet CI modes with flexible timing profiles.
