+++ ampere-cli/src/jvmTest/kotlin/link/socket/ampere/DemoCommandTest.kt
@@ -4,14 +4,14 @@ import com.github.ajalt.clikt.testing.test
 import kotlin.test.Test
 import kotlin.test.assertContains
 
-class JazzDemoCommandTest {
+class DemoCommandTest {
 
     @Test
-    fun `jazz demo help text mentions ObservabilitySpark task`() {
-        val command = JazzDemoCommand { error("context should not be needed for help output") }
+    fun `demo help text mentions ObservabilitySpark task`() {
+        val command = DemoCommand { error("context should not be needed for help output") }
         val result = command.test("--help")
 
-        assertContains(result.output, "Jazz Test demo")
+        assertContains(result.output, "AMPERE demo")
         assertContains(result.output, "ObservabilitySpark")
     }
 }
